Employees List now displaying in app, Added some API call also to help me achieve that

+++ 4thYearProjectAPI/4thYearProjectAPI/Models/Department.cs
@@ -5,8 +5,8 @@ namespace _4thYearProjectDataBaseAPI.Models
     public class Department
     {
         [Key]
-        public double deptID { get; set; }
+        public int deptID { get; set; }
         public string deptName { get; set; }
-        public double compID { get; set; }
+        public int compID { get; set; }
     }
 }
